streamlit-dashboard/requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
google-auth==2.25.2
google-api-python-client==2.111.0
